chore: [IOPID-2962]: add logout events and fix logout logic (#7031)
> [!Warning]
> This PR can be merged after Tuesday, June 3 (the day the release will
be made). This PR must necessarily be in the first beta available to
beta testers because it needs to be tested for as long as possible

> [!Warning]
> During the review of this PR you need to _**pay special attention**_
to the logic that might be related to `spawn`. I have done multiple
tests and did not seem to see any side effects, however, I encourage you
to pay special attention during the review

## Short description
This PR adds three new Mixpanel events to better track the logout flow
(start, confirm, cancel) and fixes an important issue where the logout
saga was duplicated, causing multiple logout calls and duplicated
events.

## List of changes proposed in this pull request
In this PR, I added the following three Mixpanel events to improve the
observability of the logout process:
- `LOGOUT_START`: when the user clicks “Log out from IO” in the
settings.
- `LOGOUT_CONFIRM`: when the user confirms the system alert to log out.
- `LOGOUT_CANCEL`: when the user cancels the alert and decides to stay
logged in.

Additionally, this PR includes a _**crucial fix**_:
Previously, the watchLogoutSaga was launched using a spawn in the
startupSaga. Because spawn creates independent tasks, every time the
startupSaga was re-initialized (e.g., due to session refresh or session
expiration), a new logout watcher was created without killing the old
one. This caused multiple active watchers that triggered duplicate
logout calls and duplicate Mixpanel events.

With this PR, the spawn has been replaced with a fork, ensuring that the
logout watcher is tied to the lifecycle of the startupSaga and properly
canceled when the startupSaga restarts.
As a result, the logout flow now triggers a single logout call and logs
only one event as expected.

## Demo

| Before fix ❌ | After Fix ✅ |
| - | - |
| <video
src="https://github.com/user-attachments/assets/08ff45b6-6b55-49ab-9b38-4f24c30c9ec5"/>
| <video
src="https://github.com/user-attachments/assets/a1069ee2-4522-43cd-ba90-aa70e962c28b"/>
|

## How to test
1. Run the application using dev or local environment 
2. Triggers re-startup saga **without** closing the app and reopening it
(e.g., due to session refresh or session expiration)
3. Performs the logout flow 
4. Check that on Mixpanel events are called only once and that new
events are tracked correctly

---------

Co-authored-by: Fabio Bombardi <16268789+shadowsheep1@users.noreply.github.com>