refactor: [IOPLT-1141] Create a react context to handle AppFeedback request (#7049)
## Short description
This PR refactorize the AppReview request trigger. On the first
implementation we needed to add a bottomsheet node on any usage of the
feedback flow, now the bottomsheet is generalized and the trigger is
handled by the context itself

## List of changes proposed in this pull request
- Removes previously implemented hoook
- introduced AppFeedbackContext.

## How to test
Start a new AppFeedback flow from a payment success flow or after adding
a new document on the Wallet flow.

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>