feat(IT Wallet): [SIW-2259,SIW-2473] New driving license details (#6989)
## Short description
This PR updates the style of the Driver's License credential detail at
the time L3 is enabled

## List of changes proposed in this pull request
- Change color for Driving license details header
- Added `ItwBadge`, view with a static gradient background, that
contains the it-wallet logo and a text
- With L3 enabled, restricts the generation of the trustmark to the
driver's license detail, where the sticky CTA allows users to generate
the trustmark
- Add `navigator` for ITW playground (`ItwPlaygroundStackNavigator`)


> [!NOTE]  
> The new Tooltip it's not in this PR, will be implemented in subsequent
PRs
## How to test
Verify that the new credential detail is consistent with the design on
Figma.

The new credential design is accessible from the Playground under
"Documents on IO." From there, it will be possible to navigate to the
detail screen of each credential obtained with L3 level. (Navigation to
the HEIC will only be available if you have a valid wallet; otherwise,
the navigation option will not be displayed.)



https://github.com/user-attachments/assets/1db19dea-c0db-404b-bf99-301d0844371b

---------

Co-authored-by: Cristian Matteu <94987118+ChrisMattew@users.noreply.github.com>
Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>