refactor: [IOBP-1535] CTA on CdC service screen (#7070)
## Short description
This PR refactors the CdC `useSpecialCtaCdc` hook reducing unnecessary
complexity by eliminating unused parameters and redundant logic that
this flow shouldn't provide.

## List of changes proposed in this pull request
* Removed the `serviceId` parameter and associated logic from the
`useSpecialCtaCdc` hook, including the `useServicePreferenceByChannel`
call and related state management to only check if CDC is enabled and
return the appropriate action.

## How to test
- With the dev-server started, open the service "Carta della cultura"
and check that there is a custom CTA that starts the CdC flow.

## Preview

https://github.com/user-attachments/assets/f35da1cd-c1e4-41a3-8e96-78a26a058aec