chore: [IOBP-1739] Add refetch latest receipts delay after closing the payment flow (#7080)
## Short description
This PR introduces a delay-based workaround in the
`WalletPaymentOutcomeScreen` to ensure the latest receipts are fetched
after a payment is processed. This workaround will be removed as soon as
the API to fetch the receipt details from a notice code is available.

## List of changes proposed in this pull request
* Added a `setTimeout` in the `handleClose` function to delay the
dispatch of `getPaymentsLatestReceiptAction.request()` until the payment
processing is likely complete. This ensures the receipts list is updated
when the user navigates back.

## How to test
- Start a payment flow in the UAT env.
- Complete a payment with a successful outcome, and check that after 3
seconds, the latest receipts list will refetch.