feat: [IOCOM-2152] empty message search results should be announced (#7081)
## Short description
this PR allows users to be notified by a screen reader when no messages
match the search parameters provided in the search screen

## List of changes proposed in this pull request
- required a11y i18n keys
- addition of an accessibility focusable unstyled component in the empty
message search screen, so that when navigating with an external
keyboard, an user will be able to focus it and be notified of the (lack
of) search results.

## How to test
automated tests should pass.
also check that, on both iOS and Android, when searching for messages
with a query longer than 3 characters, when:
- tapping the top half of the empty list
OR
- with the software keyboard closed, swiping right until the new
component is focused

the screen reader informs the user that no message was found.

---------

Co-authored-by: Andrea <andrea.piai@pagopa.it>