fix(IT Wallet): [SIW-2618] Obtaining eID through L3 identification (#7106)
## Short description
This PR fixes the eID retrieval when going through the L3 identification
flow.

> [!NOTE]  
> This is a temporary solution and will be replaced once the new L3
issuance flow is fully integrated

## List of changes proposed in this pull request
- Hardcoded `isL3IssuanceEnabled` to `false` when retrieving the `Wallet
Attestation`, forces usage of APIs in `io-react-native-wallet` 0.30.0.

## How to test
Ensure that the eID is correctly retrieved during the L2 issuance flow,
even when the L3 flag ( from the Documenti su IO playground) is active.