chore: [IOBP-1783] Increse TTL of the rptid saved (#7113)
## Short description
This PR updates the time-to-live (TTL) logic for cleaning expired
payments in the `cleanExpiredPaymentsOngoingFailed` saga by introducing
a new constant for TTL duration and replacing the previous reference.
This fix allows to the user to see the right screen if the
`PPT_PAGAMENTO_IN_CORSO` error persists.

## List of changes proposed in this pull request
- Introduced a new constant `PAYMENT_ONGOING_RPTID_TTL_MS` set to 24
hours (in milliseconds) to define the TTL for expired payments.

**Note**: The new constant `PAYMENT_ONGOING_RPTID_TTL_MS` is used
exclusively by the saga, which runs whenever the user opens the app for
the first time or when the session is refreshed. In contrast, the
previous constant, `PAYMENT_ONGOING_FAILURE_WAIT_TIME`, refers to the
15-minute timer that starts from the moment the user first encounters
the error.