feat(IT Wallet): [SIW-2550] Added new credential "Attestati di Laurea, Master e Dottorato" (#7100)
## Short description
This PR adds a new credential: **Attestati di Laurea, Master e
Dottorato**
This credential is not yet available for issuance and is only visible in
pre environments and with the L3 feature flag enabled.

## List of changes proposed in this pull request
- Added a new credential type: `DEGREE_CERTIFICATES`
- Created and handled a new badge: `comingSoon`
- Created the screen `ItwIssuanceCredentialComingSoonScreen` to inform
users that the credential is not yet available
- Introduced two arrays to manage credential types:
`availableCredentials` for currently issuable credentials, and
`comingSoonCredentials` for future ones (including degree certificates).
These arrays are combined and conditionally displayed through the
`displayedCredentials` variable within the `WalletCardOnboardingScreen`
component.

## How to test

1. Go to Settings → Developers → Documenti su IO and switch the
environment to Pre

1. Obtain the eID
2. Check the credentials list and verify that the new credential
"Attestati di Laurea, Master e Dottorato" is visible only in the Pre
environment and with the L3 flag enabled (Fiscal code whitelisted).
3. Tap on the credential and ensure the new screen appears "Documento in
arrivo"
4. Verify that in Prod environment or without the L3 flag, the
credential is not shown in the list.



https://github.com/user-attachments/assets/51b40a6f-e164-4a2b-afcd-e13c276d9ce4