feat: [IOBP-1796] Added `withAppRequiredUpdate` HOC to handle feature-related min app version (#7144)
## ⚠️ This PR depends on
https://github.com/pagopa/io-services-metadata/pull/998

## Short description
This PR introduces a new feature for handling app update requirements
based on specific functionalities and integrates it into existing
components. It includes a custom higher-order component (HOC)
implementation that checks whether or not the app should be updated to
access that particular feature. Currently, it's implemented on the
screens related to the `IdPay` initiative and onboarding. Additionally,
it implements new tests for the `UpdateAppAlert` component.

## List of changes proposed in this pull request

* Added a common HOC `withAppRequiredUpdate` to conditionally display
the `UpdateAppAlert` component based on app update requirements for
specific features. It uses selectors associated with a specific key to
determine if an update is needed.
* Added `idPayOnboardingRequiresAppUpdateSelector` and
`idPayDetailsRequiresAppUpdateSelector` to check if the app version
supports specific `IdPay` features.
* Refactored the screen into a base component
(`IdPayInitiativeDetailsScreenComponent`) and wrapped it with the
`withAppRequiredUpdate` HOC to enforce update checks for the
"initiative_details" feature.
* Similarly, refactored the onboarding screen into a base component
(`IdPayInitiativeDetailsScreenComponent`) and applied the HOC for the
"onboarding" feature.
* Created the `UpdateAppAlert` component to display a screen for
updating the app, including primary and secondary actions and a custom
header for support requests.
* Updated the `IO_SERVICES_METADATA_VERSION` in the
`scripts/generate-api-models.sh` file to reflect the new feature flag
for `IdPay` that must be updated as soon as the dependent PR is merged.

## How to test
- Checkout the dev-server from this PR:
https://github.com/pagopa/io-dev-api-server/pull/497
- Run `yarn generate` both on the app IO and the dev-server;
- Then, on the dev-server open the `backend.json` file and try to
increase the app version for any specific IDPay feature (such as the
`onboarding` and `initiative_details`);
- Check that if the `min_app_version` on the dev-server is higher than
the one of the IO app, the `UpdateAppAlert` component appears;

## Preview
|Onboarding|Initiative details|
|-|-|
|<video
src="https://github.com/user-attachments/assets/95857874-58d1-4b50-b0ed-0785953a71bd"
/>|<video
src="https://github.com/user-attachments/assets/316f57ba-7f6f-4bae-99a7-8d0ff0f80741"
/>