chore: [IOBP-1818] Hide IDPay timeline "show all" CTA when available only few operations (#7163)
## Short description
This PR updates the `IdPayInitiativeTimelineComponent` refactoring the
logic for conditional rendering of the "Show More" button, showing it
only if the number of transactions is greater than or equal to the max
size of the "latest transactions".

## List of changes proposed in this pull request
* Refactored the conditional rendering of the "Show More" button by
introducing a `showAllCta` constant with a typed structure. This
replaces inline logic and improves readability and maintainability.

## How to test
- Check that the "Show all" CTA is only visible when the list of
operations is more than or equal to 5.
- Simulate with the dev-server a list of only 2 operations in the list
by editing the size inside the dev-server file
(`src/persistence/idpay.ts` from line `566`)