feat: [IOBP-1816] Enable push notification at IDPay onboarding process (#7167)
## Short description
This pull request introduces a new feature to handle push notification
enablement during the IDPay onboarding process for a specific initiative

## List of changes proposed in this pull request
- Updated the `Context` type in `context.ts` to include a new property
`isPushNotificationsEnabled`
- Enhanced the state machine in `machine.ts` to include a new
`EnableNotification` state and logic to navigate to this state when
closing onboarding if notifications are not enabled
- Added a new screen `IdPayEnableNotificationScreen` in `navigator.tsx`
and updated navigation routes and parameters to include
`IDPAY_ONBOARDING_ENABLE_NOTIFICATIONS`. This screen guide users in
enabling push notifications, with actions to open system settings or
dismiss the screen.

## How to test
- Try to onboard the initiative with Id 17
- Complete the flow with push notifications enabled; onboarding should
complete normally
- Repeat onboarding with push notifications disabled, the new screen
should appear prompting the user to enable them
- After enabling it, going back to the app should navigate to `Message`
tab

## Preview


https://github.com/user-attachments/assets/e50ef156-e518-4874-a70e-97ad0aa6b28f