feat(IT Wallet): [SIW-2217,SIW-2554] Integrate CieID L3 flow inside IT-Wallet (#7165)
Depends on #7153 

## Short description
This PR enables the CieID L3 flow inside IT-Wallet with a new related
K.O. screen.

## List of changes proposed in this pull request
- Fixed the regex to support L3 only identification mode on CieID
- Added a new K.O. flow when the CieID L3 mode is not enabled on CieID
app.
- Added mixpanel tracking for ITW_CIEID_CIE_NOT_REGISTERED

## How to test

> [!IMPORTANT]
> Not testable without CieID Collaudo application and Collaudo
credentials.

Steps to test the flow:
- Have the CieID collaudo app installed (green icon) with package name
it.ipzs.cieid.coll and an account enabled for L3
- Replace any occurrence of the package name it.ipzs.cieid.collaudo with
it.ipzs.cieid.coll in app-io and reinstall the app-io Android app
- Set the isUatEnvironment parameter to true in the openCieIdApp
function

Now, proceed to activate the IT-Wallet with CieID, and everything should
work as intended.

If you want to test the new K.O. screen, remove the L3 auth on the CieID
app and retry, you should enter in the K.O. screen when trying to use
CieID as auth method.

---------

Co-authored-by: Gianluca Spada <gianluca.spada@spindox.it>