feat(IT Wallet): [SIW-2658] Extract authentication level (L2/L3) from claim verification (#7175)
This PR depends on https://github.com/pagopa/io-app/pull/7153

## Short description
This PR updates the `itwLifecycleIsITWalletValidSelector` to correctly
determine whether the credential is a valid IT-Wallet instance

## List of changes proposed in this pull request
- updated `itwLifecycleIsITWalletValidSelector`
- updated `isItwCredential` to read the claim verification and determine
the authentication level from the `assurance_level` field. If the value
is `high` then it is `L3` otherwise it is `L2`.
- removed `hasL3Design`

## How to test
- **Documenti su IO**: Disable L3, activate Documenti su IO, retrieve
all credentials and check that there are no regressions
- **IT-Wallet**: Enable L3, activate IT-Wallet with both L3 and L2,
retrieve all credentials and check that there are no regressions

---------

Co-authored-by: Gianluca Spada <gianluca.spada@spindox.it>
Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>