feat(IT Wallet): [SIW-2665] Add IT-Wallet upgrade banner in MDL details (#7161)
## Short description
This PR adds a banner in the driving license (MDL) credential details
screen to promote the IT Wallet upgrade.
The banner is shown only if the user is part of the trial system (L3
enabled) and the driving license credential is of type L2.

## List of changes proposed in this pull request
- Created the `ItwPresentationWalletUpgradeMDLDetailsBanner` component,
which represents the IT Wallet upgrade banner in the MDL credential
details screen
- Added `action`, `selectors`, and `reducer` logic to control the
visibility of the banner and allow users to dismiss it
- Added tracking for when the banner is shown, tapped, and closed.

## How to test

1. Obtain "Documenti su IO"
2. Obtain a valid "Driving license credential"
3. Verify that the new banner is displayed
4. Disable L3 from the Documenti su IO playground 
5. Verify that the new banner isn't displayed
6. Verify that the tracking events are correctly send

<img width="441" alt="Screenshot 2025-07-07 at 17 38 43"
src="https://github.com/user-attachments/assets/ad60d4e6-d835-4fa1-8e39-1a556aa81e45"
/>