chore: [IOPLT-1263] Replace `lottie-react-native` with `skottie` (`skia` library) (#7187)
## Short description
This PR replaces `lottie-react-native` with `skottie`, included in the
new version of the `react-native-skia` package.

## List of changes proposed in this pull request
- Remove `lottie-react-native`
- Update `react-native-skia` to get the new `Skottie` component
- Update `AnimatedPictogram` to use `Skottie` instead of `LottieView`
- Refactor `IngressScreen` to use the new `AnimatedPictogram` instead of
the legacy component with `LottieView`
- Add relative mock to fix Jest tests

> [!IMPORTANT]
> We attempted to add a dynamic colour to support dark mode, but the
Skottie library only supports two of the three colours (or the original
`json` file is not exported correctly), so we can't proceed.

## How to test
The performance of the animated pictogram playback should be comparable
with that of the legacy package.

---------

Co-authored-by: Emanuele Dall'Ara <71103219+LeleDallas@users.noreply.github.com>