feat: [SIW-2596] Add credential issuance flow v2 (#7194)
## Short description
This PR implements the new credential issuance flow aligned with
specification 1.0.
At this stage, due to the differences between the old and new
implementations, a temporary layer has been added to route requests to
the new flow only for enabled users. This allows the old flow to remain
active for all other users, ensuring a gradual migration to the new APIs

> [!IMPORTANT]
> This first version does not yet manage the `mso_mdoc` format 

## List of changes proposed in this pull request
- Updated the `@pagopa/io-react-native-wallet-v2` package
- Moved the content of the `itwCredentialIssuanceUtils` file into
`itwCredentialIssuanceUtilsV1` to use the former as an orchestrator
between the old and new flows
- Added `itwCredentialIssuanceUtilsV2` containing methods for the new
credential issuance flow compliant with specification 1.0
- Updated the `context` of the credential issuance state machine to
accept a list of `credentials` and the `isWhiteListed` parameter
- Updated the credential issuance machine `selectors` to reflect the
changes made to the `context`
- Updated `mapEventToFailure` to also handle potential errors from the
new issuance flow
- Modified the `onInit` action signature to update the `isWhiteListed`
key during the machine initialization phase

## How to test
### Prerequisites
Make sure you are whitelisted to test the new flow.
### New flow
Once you’ve obtained the PID (L2 or L3), request any credential you own.
The flow should appear identical to the previous one, just ensure that
the requests are targeting the new endpoints (they end with `/1-0`).

> [!NOTE]
> The status assertion step has been omitted for now, as the flow update
is still in progress
 
### Old flow
Disable the flag from **Settings -> Documents on IO**, repeat the flow,
and make sure there are no regressions and that the requests are
pointing to the old endpoints.

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>