fix(IT Wallet): [SIW-2766] delete old eID when obtaining a new one (#7217)
## Short description
This PR removes the old eID when a new one is obtained. This should not
be an issue normally, but when upgrading from Documenti su IO to
IT-Wallet the credentials store will contain both credentials
(`PersonIdentificationData` and `dc_sd_jwt_PersonIdentificationData`),
because their ID is different.

It also solves another problem: when deleting a credential all
credentials of the same type are removed from the store, but not all of
their cryptographic keys are removed from the device. This might happen
for a MDL that is stored in both SD-JWT and mDoc format.

## List of changes proposed in this pull request
- Modified the action to remove a credential to take a credential type
- Modified `handleItwCredentialsRemoveSaga` to select all credentials of
that type, and then remove them from the store along with their keys
- Dispatched the action to remove the old eID before storing the new one

## How to test
Activate Documenti su IO, then upgrade to IT-Wallet.

- **Before:** credentials reducer contains `PersonIdentificationData`
and `dc_sd_jwt_PersonIdentificationData`
- **After:** credentials reducer only contains
`dc_sd_jwt_PersonIdentificationData`