feat(IT Wallet): [SIW-2559] Include Wallet Attestation in remote presentation (#7233)
## Short description
This PR adds the Wallet Attestation in SD-JWT format in the remote
presentation. The Wallet Attestation is included behind the scenes and
not shown to the user.

If it is expired or invalid, it is fetched and stored in the
presentation machine.

## List of changes proposed in this pull request
- Added states to check and request the Wallet Attestation in the
presentation machine
- Included the attestation in the `getPresentationDetails` actor

## How to test
Start a presentation requesting the PID and the Wallet Attestation using
the test RP. Ensure that
- The presentation succeeds
- The Wallet Attestation claims are hidden in the Wallet's consent
screen
- The Wallet Attestation claims are visible in the RP's recap screen