fix(IT Wallet): [SIW-2463] Mixpanel credential status events (#7069)
## Short description
This PR fixes how credential status is tracked in Mixpanel in three key
areas:

1. When updating profile properties

1. When updating super properties

1. When dispatching the `ITW_CREDENTIAL_DETAIL` event

## List of changes proposed in this pull request
- Introduced the `getCredentialMixpanelStatus` utility to consistently
map a credential (when present) to the correct Mixpanel status. This
function is now used by handlers to compute the status of specific
credentials (MDL, TS, CED) when updating profile and super properties.
- Added `mapEidStatusToMixpanel`, a function for mapping the eID
credential status to the appropriate Mixpanel status
- Replaced the `ID_STATUS_MAP` with `CREDENTIALS_STATUS_MAP`, which
provides a mapping of `getCredentialStatus` values to Mixpanel statuses
( this map is not used for the eID, since its status is handled
separately via `mapEidStatusToMixpanel`)

> [!IMPORTANT]  
> This change does not take in consideration L3 credentials and PID. 

## How to test

1. Mock the `getCredentialStatus` return value for each credential type
1. Send the `ITW_CREDENTIAL_DETAIL` event and verify that the correct
credential status is sent to Mixpanel
1. Check that the Mixpanel profile properties are updated with the
expected credential status values.

1. Check that the Mixpanel super properties are updated with the
expected credential status values.