fix(IT Wallet): [SIW-2813] Split attestation utils between v1 and v2 (#7235)
## Short description
This PR splits `itwAttestationUtils.ts` into two files to avoid mixed
imports of `io-react-native-wallet` v1 and v2.

Without this clear separation, `createCryptoContextFor` was imported
only from v1, so the Wallet Attestation's `cnf.jwk` could be affected by
the 33 bytes issue; later, when using the Wallet Attestation to request
a IT-Wallet credential, the `createCryptoContextFor` from v2 was used
and the two JWKs did not match.

> [!NOTE]
> The separation will be removed alongside the two versions
`io-react-native-wallet`.

## List of changes proposed in this pull request
- Split `itwAttestationUtils.ts` into `itwAttestationUtils.v1.ts` and
`itwAttestationUtils.v2.ts`
- Used the whitelisted status to choose between v1 and v2

## How to test
The problem is not easy to reproduce, as it only happens when a JWK with
a coordinate of 33 bytes is generated. However, ensure the Wallet
Attestation is fetched correctly when L3 flag is enabled and when it is
disabled.