feat(IT Wallet): [SIW-2819] Adds access counter for offline wallet (#7241)
## Short description
This PR implements an access counter for the offline mini-app. 
The counter is increased every time an user completes the identification
(pin or biometric) and opens the offline mini-app.
The counter is reset when the app manages to update the wallet instance
status successfully

## List of changes proposed in this pull request
- Added `securePreferences` store, persisted in the secure storage
- Added sagas to handle counter increase and reset

## How to test
- Open the offline mini app
- With the help of Reactotron or similar tools, check that the counter
is increased every time the offline mini-app is displayed
- Repeat the above steps many times and check that the counter increases
correctly
- Put the app in the background and wait 120 seconds, return to the app
and check that the counter is increased
- Switch to the online version of app
- Check that the counter is reset to 0 as soon as the identification is
completed